# docs/design/api.yaml
openapi: 3.0.3
info:
  title: 勤怠管理システム API
  version: 1.0.0
  description: 社員の出退勤打刻、勤怠管理、休暇申請を管理する REST API
servers:
  - url: http://localhost:8080/api
    description: ローカル開発
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: ログイン
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: ログインユーザー情報取得
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'

  # === 打刻 ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [employeeCode, password]
      properties:
        employeeCode:
          type: string
          example: "tanaka001"
        password:
          type: string
          format: password
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: '#/components/schemas/EmployeeResponse'
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeCode:
          type: string
        name:
          type: string
        email:
          type: string
        departmentName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
...
    Role:
      type: string
      enum: [EMPLOYEE, MANAGER]
...